Decouple the UUID generator from network interfaces by having MAC addresses added to the UUID generator using uuid_ether_add(). The UUID generator keeps an arbitrary number of MAC addresses, under the assumption that they are rarely removed (= uuid_ether_del()). This achieves the following: 1.  It brings up closer to having the network stack as a loadable     module. 2.  It allows the UUID generator to filter MAC addresses for best     results (= highest chance of uniqeness). 3.  MAC addresses can come from anywhere, irrespactive of whether     it's used for an interface or not.
A side-effect of the change is that when no MAC addresses have been
added, a random multicast MAC address is created once and re-used if
needed. Previusly, when a random MAC address was needed, it was
created for every call. Thus, a change in behaviour is introduced
for when no MAC addresses exist.

Obtained from:	Juniper Networks, Inc.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@253590 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f