Add a fix for Long-standing problem with VMware. Described in below links: https://communities.vmware.com/thread/107230 https://communities.vmware.com/docs/DOC-11677
Basically, ignore the ``function 62'' and ``function 63'' interpretations
of the left/right command key when we're in the lengthiest portion of the
installation (initiated by the `auto' module).

The net effect is that you can now (once you've started the installer from
the media) escape the VM without prematurely terminating the current action
due to spurious escape sequence.

MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@259468 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f